Extract -src crate for wolfssl
