+++ crates/wolfssl-src/patches/fix-CVE-2022-25640.patch
