+++ crates/wolfssl-src/patches/fix-CVE-2022-39173.patch
