+++ crates/wolfssl-src/patches/fix-CVE-2022-42905.patch
